+++ eras/shelley/impl/test/Main.hs
@@ -1,10 +1,10 @@
 module Main where
 
 import Test.Cardano.Ledger.Common
-import qualified Test.Cardano.Ledger.Shelley.Binary.GoldenSpec as GoldenSpec
+import qualified Test.Cardano.Ledger.Shelley.BinarySpec as Binary
 
 main :: IO ()
 main =
   ledgerTestMain $
     describe "Shelley" $ do
-      GoldenSpec.spec
+      Binary.spec