+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Arbitrary.hs
@@ -30,7 +30,6 @@ import qualified Cardano.Crypto.VRF as VRF
 import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (EncCBOR)
 import Cardano.Ledger.Crypto
-import Cardano.Ledger.EpochBoundary
 import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.API (
   ApplyTxError (ApplyTxError),
@@ -151,24 +150,6 @@ instance
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
-  shrink = genericShrink
-
-instance Crypto c => Arbitrary (SnapShot c) where
-  arbitrary =
-    SnapShot
-      <$> arbitrary
-      <*> arbitrary
-      <*> arbitrary
-  shrink = genericShrink
-
-instance Crypto c => Arbitrary (SnapShots c) where
-  arbitrary = do
-    ssStakeMark <- arbitrary
-    ssStakeSet <- arbitrary
-    ssStakeGo <- arbitrary
-    ssFee <- arbitrary
-    let ssStakeMarkPoolDistr = calculatePoolDistr ssStakeMark
-    pure $ SnapShots {..}
   shrink = genericShrink
 
 instance
@@ -318,7 +299,7 @@ instance Crypto c => Arbitrary (FreeVars c) where
   shrink = genericShrink
 
 ------------------------------------------------------------------------------------------
--- Cardano.Ledger.Shelley.Rules ----------------------------------------------------------
+-- Cardano.Ledger.Shelley.Governance -----------------------------------------------------
 ------------------------------------------------------------------------------------------
 
 instance
@@ -331,8 +312,11 @@ instance
   arbitrary = ShelleyGovState <$> arbitrary <*> arbitrary <*> arbitrary <*> arbitrary
   shrink = genericShrink
 
+instance Era era => Arbitrary (Constitution era) where
+  arbitrary = Constitution <$> arbitrary <*> arbitrary
+
 ------------------------------------------------------------------------------------------
--- Cardano.Ledger.Shelley.TxAuxData ----------------------------------------------------------
+-- Cardano.Ledger.Shelley.TxAuxData ------------------------------------------------------
 ------------------------------------------------------------------------------------------
 
 -- | Max size of generated Metadatum List and Map